requirements.txt
dlt[bigquery]>=0.3.24
google-auth==2.3.3
google-cloud-secret-manager==2.7.0
requests==2.26.0
dlt
slackclient==2.9.4